+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Spawn camera and spaceship player sprite
 - Basic player controls
+- Rapier physics engine for ship controls
 - FixedTime and FixedUpdate for game systems
 - Game assets from kenney.nl
 - Git LFS config for image storage